feat: moved tests into test_haplotype.py
