+++ divref/tests/test_haplotype.py
@@ -1,6 +1,8 @@
 """Tests for shared Hail utilities in haplotype.py."""
 
 from pathlib import Path
+from typing import Any
+from unittest.mock import patch
 
 import hail as hl
 import pytest
@@ -15,8 +17,8 @@ from divref.haplotype import variant_distance
 # ---------------------------------------------------------------------------
 
 
-def _make_variant(position: int, ref: str, alt: str) -> hl.Struct:
-    return hl.Struct(locus=hl.Struct(contig="chr1", position=position), alleles=[ref, alt])
+def _make_variant(position: int, ref: str, alt: str, contig: str = "chr1") -> hl.Struct:
+    return hl.Struct(locus=hl.Struct(contig=contig, position=position), alleles=[ref, alt])
 
 
 def _make_haplotype_table(variant_positions: list[tuple[str, int, str, str]]) -> hl.Table:
@@ -47,6 +49,48 @@ def _make_haplotype_table(variant_positions: list[tuple[str, int, str, str]]) ->
 # ---------------------------------------------------------------------------
 # get_haplo_sequence
 # ---------------------------------------------------------------------------
+
+
+def _create_reference_mock(reference_sequence: str) -> Any:
+    """
+    Create a mock for hl.get_sequence that returns substrings of a fixed reference.
+
+    The mock accepts the same arguments as hl.get_sequence and returns the
+    appropriate substring of the provided reference string.
+
+    Args:
+        reference_sequence: The reference string to use for subsequence extraction.
+
+    Returns:
+        A callable that mimics hl.get_sequence using the provided reference.
+    """
+
+    def mock_get_sequence(
+        _contig: str,
+        position: int,
+        before: int = 0,
+        after: int = 0,
+        reference_genome: Any = None,  # noqa: ARG001
+    ) -> Any:
+        return hl.str(reference_sequence)[position - before : position + after + 1]
+
+    return mock_get_sequence
+
+
+def test_get_haplo_sequence_edge_cases(hail_context: None) -> None:  # noqa: ARG001
+    """Test get_haplo_sequence with SNPs, insertions, and deletions."""
+    reference = "01234567891"
+
+    two_snps = [_make_variant(4, "A", "T"), _make_variant(6, "G", "C")]
+    two_insertions = [_make_variant(4, "A", "AT"), _make_variant(6, "G", "GC")]
+    two_deletions = [_make_variant(4, "AT", "A"), _make_variant(7, "GC", "G")]
+
+    mock_get_sequence = _create_reference_mock(reference)
+
+    with patch("hail.get_sequence", side_effect=mock_get_sequence):
+        assert hl.eval(get_haplo_sequence(context_size=2, variants=two_snps)) == "23T5C78"
+        assert hl.eval(get_haplo_sequence(context_size=2, variants=two_insertions)) == "23AT5GC78"
+        assert hl.eval(get_haplo_sequence(context_size=2, variants=two_deletions)) == "23A6G91"
 
 
 def test_get_haplo_sequence_single(
